Завершил дз для 5 семинара.

+++ DZ_Seminar_5/src/model/ProdModel.java
@@ -3,18 +3,18 @@ package model;
 public class ProdModel extends CalcModel {
 
     @Override
-    public int result() {
+    public double result() {
         return x * y;
     }
 
     @Override
-    public void setX(int value) {
+    public void setX(double value) {
         super.x = value;
 
     }
 
     @Override
-    public void setY(int value) {
+    public void setY(double value) {
         super.y = value;
 
     }
