+++ DZ_Seminar_5/src/model/SumModel.java
@@ -6,18 +6,18 @@ public class SumModel extends CalcModel {
 
 
     @Override
-    public int result() {
+    public double result() {
         return x + y;
     }
 
     @Override
-    public void setX(int value) {
+    public void setX(double value) {
         super.x = value;
 
     }
 
     @Override
-    public void setY(int value) {
+    public void setY(double value) {
         super.y = value;
 
     }
